+++ Assets/Asset Packs/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5b4f1a838db6e64ba1808385a89555d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
