+++ Assets/Game/Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c3e17dcdaf828004ea72a29a294885c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
