+++ Assets/Game/Characters/Character.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54f38660fa6971f49ba438aa17d0849c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
