+++ Assets/Game/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3555f368f58d109408e1b6e417dc8644
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
