+++ Assets/Game/Core/CameraFollow.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 5ef513b621f0bd04eb9df35f2e605227
-folderAsset: yes
-DefaultImporter:
+guid: 1f80192e74fc2d74b9da3ea5ee9e7d90
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
