+++ Assets/Game/Enviroments.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3b958b3c66bdc844982811d6fb44ecb4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
