+++ Assets/Game/Enviroments/DarkBrawnLayer.terrainlayer
@@ -6,8 +6,8 @@ TerrainLayer:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: NewLayer 2
-  m_DiffuseTexture: {fileID: 2800000, guid: 8b6712a039f2cf24fb0cb22f5be29af8, type: 3}
+  m_Name: DarkBrawnLayer
+  m_DiffuseTexture: {fileID: 2800000, guid: 3c31f7cb4de039848a73c7f61c59707f, type: 3}
   m_NormalMapTexture: {fileID: 0}
   m_MaskMapTexture: {fileID: 0}
   m_TileSize: {x: 2, y: 2}
