+++ Assets/Game/Enviroments/DarkGreenLayer.terrainlayer
@@ -6,7 +6,7 @@ TerrainLayer:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: NewLayer
+  m_Name: DarkGreenLayer
   m_DiffuseTexture: {fileID: 2800000, guid: 97fbaf45bf3e43b45ba4cb8c6873bb51, type: 3}
   m_NormalMapTexture: {fileID: 0}
   m_MaskMapTexture: {fileID: 0}
