+++ Assets/Game/Enviroments/LightGreenLayer.terrainlayer
@@ -6,8 +6,8 @@ TerrainLayer:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: NewLayer 1
-  m_DiffuseTexture: {fileID: 2800000, guid: 80598239e7a383e4b8472b6305194c22, type: 3}
+  m_Name: LightGreenLayer
+  m_DiffuseTexture: {fileID: 2800000, guid: 8b6712a039f2cf24fb0cb22f5be29af8, type: 3}
   m_NormalMapTexture: {fileID: 0}
   m_MaskMapTexture: {fileID: 0}
   m_TileSize: {x: 2, y: 2}
