--- Assets/Scripts/Movement.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using UnityEditor;
-using UnityEngine;
-using UnityEngine.AI;
-
-public class Movement : MonoBehaviour {
-    void Update() {
-        if (Input.GetMouseButtonDown(0)) {
-            MoveToCursor();
-        }
-    }
-
-    private void MoveToCursor() {
-        Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
-        RaycastHit hit;
-        bool hasHit = Physics.Raycast(ray, out hit);
-        if (hasHit) {
-            GetComponent<NavMeshAgent>().destination = hit.point;
-        }
-    }
-}
